==> 14.explore_wallets.ipynb <==
### Wallet pairs exploration

# Read datasets and merge
df = pd.read_csv('data/data_features.csv')
df1 = pd.read_csv('data/data_sybils_pairs.csv')
df1['target'] = 1

df = df.merge(df1, how = 'left')
df['target'] = df['target'].fillna(0).astype(int)

print(df.shape)
df[:3]

len_wal = lambda x: len(set('_'.join(x).split('_')))

dict_sybils = df1.groupby('round_name')['wallets'].agg(len_wal).to_dict()
dict_sybils

### 1. Statistics

# Basic stat
df2 = pd.read_csv('data/data_rounds.csv')
df2 = df2.groupby('round_name').agg({
    'source_wallet': lambda x: len(set(x)),
    # 'destination_wallet': lambda x: len(set(x)),
    # 'amt_usd': [len, sum],    
})
df2.columns = [
    'Number of donors',
    # 'Number of projects',
    # 'Number of votes',
    # 'Donated sum'
]
# df1['Average number of votes per donor'] = df1['Number of votes']/df1['Number of donors']
# df1['Average sum per donor'] = df1['Donated sum']/df1['Number of donors']
df2.reset_index(inplace = True)

# Adding stat
f1 = df.tenure_round.notnull()
f2 = df.target == 1

dict1 = df.groupby('round_name')['wallets'].agg(len_wal).to_dict()
df2['Connected donors per year'] = df2.round_name.map(dict1)

dict1 = df[f1].groupby('round_name')['wallets'].agg(len_wal).to_dict()
df2['Connected donors per round'] = df2.round_name.map(dict1)

df2['Sybils by hash algorithm'] = df2.round_name.map(dict_sybils)

dict1 = df[f2].groupby('round_name')['wallets'].agg(len_wal).to_dict()
df2['Sybils approved by trnx per year'] = df2.round_name.map(dict1)

dict1 = df[f1&f2].groupby('round_name')['wallets'].agg(len_wal).to_dict()
df2['Sybils approved by trnx per round'] = df2.round_name.map(dict1)

df2

### 2. Generate output file with potential Sybil wallets

# Hash algo
df1 = pd.read_csv('data/data_sybils.csv')
df1 = df1.groupby(['feat_round','source_wallet']).size().reset_index()
del df1[0]
df1.columns = ['round_name','wallet']
df1['flg_hash_algo'] = 1

# Trnx verification
wal = lambda x: set('_'.join(x).split('_'))
f1 = df.tenure_round.notnull()
f2 = df.target == 1

###
list1 = list()
list2 = list()
x = df[f2].groupby('round_name')['wallets'].agg(wal)
for ind,x1 in x.iteritems():
    list1+=list(x1)
    list2+=[ind]*len(x1)
df0 = pd.DataFrame({
    'round_name': list2,
    'wallet': list1
})
feat = 'flg_trnx_year'
df0[feat] = 1
df1 = df1.merge(df0, how = 'left')
df1[feat] = df1[feat].fillna(0).astype(int)

###
list1 = list()
list2 = list()
x = df[f2&f1].groupby('round_name')['wallets'].agg(wal)
for ind,x1 in x.iteritems():
    list1+=list(x1)
    list2+=[ind]*len(x1)
df0 = pd.DataFrame({
    'round_name': list2,
    'wallet': list1
})
feat = 'flg_trnx_round'
df0[feat] = 1
df1 = df1.merge(df0, how = 'left')
df1[feat] = df1[feat].fillna(0).astype(int)

df1.to_csv('data/wallets_sybils.csv', index = False)
print(df1.shape)
df1[:5]

df1.groupby('round_name').sum()
